Create main.py
Initial upload of text entry for SSID & Password

Up: Increase letter / number
Down: Decrease letter / number
Left: Move cursor to the left
Right: Move cursor to the right
Click: Add a new character
A: Confirm the SSID or Password, if on the password line, try connecting
B: Switch letter to lowercase, if the cursor is on the last letter and the letter is lowercase already, delete. If on a number or symbol, do nothing.